+++ week4/MoonSeoIn/src/components/FixedButton.jsx
@@ -1,0 +1,9 @@
+const FixedButton = () => {
+  return (
+    <button className="fixed-btn">
+      <img src="./plus.svg" alt="글쓰기 버튼" />
+    </button>
+  );
+};
+
+export default FixedButton;
